requirements.txt
kivy==2.3.1
requests==2.31.0
arabic-reshaper==2.1.3
python-bidi==0.4.2
